requirements.txt
requests
beautifulsoup4
Flask
Flask-Cors
gunicorn
psycopg2-binary
python-dotenv
